--- model/staff.js
@@ -1,4 +1,0 @@
-var mongoose = require('mongoose');
-var articleSchema = require('./../schema/staff.js');
-const articleModel = mongoose.model('Article',articleSchema);
-module.exports =  articleModel;
